requirements.txt
streamlit
openai
python-dotenv

langchain
langchain-community
langchain-text-splitters

faiss-cpu
sentence-transformers
transformers